Removed unnecessary tests. Fixed some style.
