+++ src/test/java/seedu/module/model/module/NameContainsKeywordsPredicateTest.java
@@ -1,78 +1,69 @@
-// package seedu.module.model.module;
+package seedu.module.model.module;
 
-// import static org.junit.jupiter.api.Assertions.assertFalse;
-// import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-// import java.util.Arrays;
-// import java.util.Collections;
-// import java.util.List;
+import java.util.ArrayList;
 
-// import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Test;
 
-// import seedu.module.testutil.PersonBuilder;
+import seedu.module.model.module.predicate.NameContainsKeywordsPredicate;
+import seedu.module.testutil.ArchivedModuleBuilder;
 
-// public class NameContainsKeywordsPredicateTest {
+public class NameContainsKeywordsPredicateTest {
 
-//     @Test
-//     public void equals() {
-//         List<String> firstPredicateKeywordList = Collections.singletonList("first");
-//         List<String> secondPredicateKeywordList = Arrays.asList("first", "second");
+    ArchivedModule archivedModule = new ArchivedModuleBuilder().build();
 
-//         NameContainsKeywordsPredicate firstPredicate = new NameContainsKeywordsPredicate(firstPredicateKeywordList);
-//         NameContainsKeywordsPredicate secondPredicate = new NameContainsKeywordsPredicate(
-//              secondPredicateKeywordList);
 
-//         // same object -> returns true
-//         assertTrue(firstPredicate.equals(firstPredicate));
+    @Test
+    public void test_SameModuleCode_returnsTrue() {
+        ArrayList<String> list = new ArrayList<>() {
+            {
+                add("CS2103T");
+            }
+        };
+        NameContainsKeywordsPredicate nameContainsKeywordsPredicate = new NameContainsKeywordsPredicate(list);
+        assertTrue(nameContainsKeywordsPredicate.test(archivedModule));
+    }
 
-//         // same values -> returns true
-//         NameContainsKeywordsPredicate firstPredicateCopy = new NameContainsKeywordsPredicate(
-//             firstPredicateKeywordList);
-//         assertTrue(firstPredicate.equals(firstPredicateCopy));
+    @Test
+    public void test_SameModuleCodeLowerCase_returnsTrue() {
+        ArrayList<String> list = new ArrayList<>() {
+            {
+                add("cs2103t");
+            }
+        };
+        NameContainsKeywordsPredicate nameContainsKeywordsPredicate = new NameContainsKeywordsPredicate(list);
+        assertTrue(nameContainsKeywordsPredicate.test(archivedModule));
+    }
 
-//         // different types -> returns false
-//         assertFalse(firstPredicate.equals(1));
+    @Test
+    public void test_differentModuleCode_returnsFalse() {
+        ArrayList<String> list = new ArrayList<>() {
+            {
+                add("cs2101");
+            }
+        };
+        NameContainsKeywordsPredicate nameContainsKeywordsPredicate = new NameContainsKeywordsPredicate(list);
+        assertFalse(nameContainsKeywordsPredicate.test(archivedModule));
+    }
 
-//         // null -> returns false
-//         assertFalse(firstPredicate.equals(null));
+    @Test
+    public void test_partialModuleCode_returnsFalse() {
+        ArrayList<String> list = new ArrayList<>() {
+            {
+                add("CS2103");
+            }
+        };
+        NameContainsKeywordsPredicate nameContainsKeywordsPredicate = new NameContainsKeywordsPredicate(list);
+        assertFalse(nameContainsKeywordsPredicate.test(archivedModule));
+    }
 
-//         // different person -> returns false
-//         assertFalse(firstPredicate.equals(secondPredicate));
-//     }
+    @Test
+    public void test_emptyList_returnsFalse() {
+        ArrayList<String> list = new ArrayList<>();
+        NameContainsKeywordsPredicate nameContainsKeywordsPredicate = new NameContainsKeywordsPredicate(list);
+        assertFalse(nameContainsKeywordsPredicate.test(archivedModule));
+    }
 
-//     @Test
-//     public void test_nameContainsKeywords_returnsTrue() {
-//         // One keyword
-//         NameContainsKeywordsPredicate predicate = new NameContainsKeywordsPredicate(
-//             Collections.singletonList("Alice"));
-//         assertTrue(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
-
-//         // Multiple keywords
-//         predicate = new NameContainsKeywordsPredicate(Arrays.asList("Alice", "Bob"));
-//         assertTrue(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
-
-//         // Only one matching keyword
-//         predicate = new NameContainsKeywordsPredicate(Arrays.asList("Bob", "Carol"));
-//         assertTrue(predicate.test(new PersonBuilder().withName("Alice Carol").build()));
-
-//         // Mixed-case keywords
-//         predicate = new NameContainsKeywordsPredicate(Arrays.asList("aLIce", "bOB"));
-//         assertTrue(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
-//     }
-
-//     @Test
-//     public void test_nameDoesNotContainKeywords_returnsFalse() {
-//         // Zero keywords
-//         NameContainsKeywordsPredicate predicate = new NameContainsKeywordsPredicate(Collections.emptyList());
-//         assertFalse(predicate.test(new PersonBuilder().withName("Alice").build()));
-
-//         // Non-matching keyword
-//         predicate = new NameContainsKeywordsPredicate(Arrays.asList("Carol"));
-//         assertFalse(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
-
-//         // Keywords match phone, email and address, but does not match name
-//         predicate = new NameContainsKeywordsPredicate(Arrays.asList("12345", "alice@email.com", "Main", "Street"));
-//         assertFalse(predicate.test(new PersonBuilder().withName("Alice").withPhone("12345")
-//                 .withEmail("alice@email.com").withAddress("Main Street").build()));
-//     }
-// }
+}
